+++ usr/include/grp.h
@@ -1,6 +1,6 @@
 struct	group { /* see getgrent(3) */
 	char	*gr_name;
 	char	*gr_passwd;
-	int	gr_gid;
+	short	gr_gid;
 	char	**gr_mem;
 };
